COMP: Fix building against the SlicerOpenCV extension
After feedback, renamed the SLICEROPENCV_DIR to SlicerOpenCV_DIR and
redefined it as the build directory, as can use the Slicer defined path to
the Qt loadable module library dir to build the include and lib paths.
Added missing opencv libraries to link against.
Add the version number to the end of the ITKVideoBridgeOpenCV library as
the currently produced library file doesn't have a non versioned symlink
to the versioned library file.